added cast

+++ include/boost/unordered/detail/foa.hpp
@@ -1559,7 +1559,7 @@ private:
         prefetch_elements(p);
         do{
           //auto n=unchecked_countr_zero(mask);
-          auto n=__builtin_ctz((unsigned int)mask);
+          auto n=(unsigned int)__builtin_ctz((unsigned int)mask);
           if(BOOST_LIKELY(bool(pred()(x,key_from(p[n]))))){
             return {pg,n,p+n};
           }
